refactor: Move related files logic to store
