+++ source/types/renderer/misc.ts
@@ -1,0 +1,6 @@
+export interface RelatedFile {
+  file: string
+  path: string
+  tags: string[]
+  link: 'inbound'|'outbound'|'bidirectional'|'none'
+}
