+++ source/win-main/MainSidebar.vue
@@ -160,19 +160,13 @@ import { DirMeta, MDFileMeta, OtherFileMeta } from '@dts/common/fsal'
 import { TabbarControl } from '@dts/renderer/window'
 import sanitizeHtml from 'sanitize-html'
 import { getConverter } from '@common/util/md-to-html'
+import { RelatedFile } from '@dts/renderer/misc'
 
 const path = window.path
 const ipcRenderer = window.ipc
 
 // Must be instantiated after loading, i.e. when the Sidebar is initialized
 let md2html: Function
-
-interface RelatedFile {
-  file: string
-  path: string
-  tags: string[]
-  link: 'inbound'|'outbound'|'bidirectional'|'none'
-}
 
 export default defineComponent({
   name: 'MainSidebar',
@@ -181,8 +175,7 @@ export default defineComponent({
   },
   data: function () {
     return {
-      bibContents: undefined as undefined|any[],
-      relatedFiles: [] as RelatedFile[]
+      bibContents: undefined as undefined|any[]
     }
   },
   computed: {
@@ -251,6 +244,9 @@ export default defineComponent({
     activeFile: function (): MDFileMeta|null {
       return this.$store.state.activeFile
     },
+    relatedFiles: function (): RelatedFile[] {
+      return this.$store.state.relatedFiles
+    },
     modifiedFiles: function (): string[] {
       return this.$store.state.modifiedDocuments
     },
@@ -290,9 +286,6 @@ export default defineComponent({
       // Reload the bibliography
       this.updateReferences().catch(e => console.error('Could not update references', e))
     },
-    activeFile: function () {
-      this.updateRelatedFiles().catch(e => console.error('Could not update related files', e))
-    },
     modifiedFiles: function () {
       if (this.activeFile == null) {
         return
@@ -302,7 +295,8 @@ export default defineComponent({
       // immediately account for any changes in the related files.
       const activePath = this.activeFile.path
       if (!(activePath in this.modifiedFiles)) {
-        this.updateRelatedFiles().catch(e => console.error('Could not update related files', e))
+        this.$store.dispatch('updateRelatedFiles')
+          .catch(e => console.error('Could not update related files', e))
       }
     }
   },
@@ -314,7 +308,8 @@ export default defineComponent({
     })
 
     ipcRenderer.on('links', () => {
-      this.updateRelatedFiles().catch(err => console.error(err))
+      this.$store.dispatch('updateRelatedFiles')
+        .catch(e => console.error('Could not update related files', e))
     })
 
     try {
@@ -322,7 +317,6 @@ export default defineComponent({
     } catch (err) {
       console.error(err)
     }
-    this.updateRelatedFiles().catch(e => console.error('Could not update related files', e))
   },
   created: function () {
     // Instantiate a converter so that we can convert the md of our ToC entries
@@ -340,87 +334,6 @@ export default defineComponent({
         command: 'get-bibliography',
         payload: this.citationKeys.map(e => e)
       })
-    },
-    updateRelatedFiles: async function () {
-      // First reset, default is no related files
-      this.relatedFiles = []
-      if (this.activeFile === null || this.activeFile.type !== 'file') {
-        return
-      }
-
-      const unreactiveList: RelatedFile[] = []
-
-      // Then retrieve the inbound links first, since that is the most important
-      // relation, so they should be on top of the list.
-      const { inbound, outbound } = await ipcRenderer.invoke('link-provider', {
-        command: 'get-inbound-links',
-        payload: { filePath: this.activeFile.path }
-      }) as { inbound: string[], outbound: string[]}
-
-      for (const absPath of [ ...inbound, ...outbound ]) {
-        const found = unreactiveList.find(elem => elem.path === absPath)
-        if (found !== undefined) {
-          continue
-        }
-
-        const related: RelatedFile = {
-          file: path.basename(absPath),
-          path: absPath,
-          tags: [],
-          link: 'none'
-        }
-
-        if (inbound.includes(absPath) && outbound.includes(absPath)) {
-          related.link = 'bidirectional'
-        } else if (inbound.includes(absPath)) {
-          related.link = 'inbound'
-        } else {
-          related.link = 'outbound'
-        }
-
-        unreactiveList.push(related)
-      }
-
-      // The second way files can be related to each other is via shared tags.
-      // This relation is not as important as explicit links, so they should
-      // be below the inbound linked files.
-      const recommendations = await ipcRenderer.invoke('tag-provider', {
-        command: 'recommend-matching-files',
-        payload: this.activeFile.tags.map(tag => tag) // De-proxy
-      })
-
-      // Recommendations come in the form of [file: string]: string[]
-      for (const filePath of Object.keys(recommendations)) {
-        const existingFile = unreactiveList.find(elem => elem.path === filePath)
-        if (existingFile !== undefined) {
-          // This file already links here
-          existingFile.tags = recommendations[filePath]
-        } else {
-          // This file doesn't explicitly link here but it shares tags
-          unreactiveList.push({
-            file: path.basename(filePath),
-            path: filePath,
-            tags: recommendations[filePath],
-            link: 'none'
-          })
-        }
-      }
-
-      // Now we have all relations based on either tags or backlinks. We must
-      // now order them in such a way that the hierarchy is like that:
-      // 1. Backlinks that also share common tags
-      // 2. Backlinks that do not share common tags
-      // 3. Files that only share common tags
-      const backlinksAndTags = unreactiveList.filter(e => e.link !== 'none' && e.tags.length > 0)
-      backlinksAndTags.sort((a, b) => { return b.tags.length - a.tags.length })
-
-      const backlinksOnly = unreactiveList.filter(e => e.link !== 'none' && e.tags.length === 0)
-      // No sorting necessary
-
-      const tagsOnly = unreactiveList.filter(e => e.link === 'none')
-      tagsOnly.sort((a, b) => { return b.tags.length - a.tags.length })
-
-      this.relatedFiles = [ ...backlinksAndTags, ...backlinksOnly, ...tagsOnly ]
     },
     getIcon: function (attachmentPath: string) {
       const fileExtIcon = ClarityIcons.get('file-ext')
@@ -452,7 +365,7 @@ export default defineComponent({
     },
     getRelatedFileName: function (filePath: string) {
       const descriptor = this.$store.getters.file(filePath)
-      if (descriptor === null) {
+      if (descriptor === undefined) {
         return filePath
       }
